fix(animation): progressEnd coercion is reset before onFinish (#28394)
Issue number: resolves #28393

---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

Our animation library's value coercion is not reset before developer
`onFinish` callbacks fire. This can lead to developers getting incorrect
state when querying for `getDuration` or `getDirection`

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Internal value coercion is reset before developer `onFinish` callbacks
fire.

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

I'm putting this in a minor release to minimize risk. This is not a
breaking change, but there may be developers relying on this broken
behavior to implement a workaround.

<!-- If this introduces a breaking change, please describe the impact
and migration path for existing applications below. -->


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->

Note: This change is needed for the toast swipe to dismiss feature
(FW-2004)